[ci] unpin workflows from master
We initially pinned workflows to master to help with forwards compatibility (so that old PRs based against master would continue to work).

However, in the long run we want to make the workflows reflect whatever is currently in the local check out, which makes it easier to test changes to workflows.

Since it's been a week since the workflow consolidation: doing the unpinning now. Will follow up with a PR to give a nice message for any PRs that are left stale.

ghstack-source-id: c8abd4564a430b2eae853b12fd4e6916fb7d82bb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74817